fix linux build break

+++ src/perf/lib/ThroughputClient.cpp
@@ -587,8 +587,8 @@ ThroughputClient::StreamCallback(
             uint32_t BufferLength = sizeof(Timings);
             MsQuic->GetParam(StreamHandle, QUIC_PARAM_STREAM_BLOCKED_TIMINGS, &BufferLength, Timings);
             WriteOutput("Flow blocked timing:\n");
-            for (int i = 0; i < ARRAYSIZE(Timings); ++i) {
-                WriteOutput("Reason: %d Time: %llu\n", Timings[i].Reason, Timings[i].Time);
+            for (size_t i = 0; i < ARRAYSIZE(Timings); ++i) {
+                WriteOutput("Reason: %d Time: %llu\n", Timings[i].Reason, (unsigned long long)Timings[i].Time);
             }
         }
         OnStreamShutdownComplete(StrmContext);
